pyproject.toml
[tool.poetry]
name = "tottle-types"
version = "2.0.2"
description = ""
authors = ["exthrempty <andrewsalt.e@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.1"
pyhumps = "^3.0.2"
pydantic = "^1.8.2"
loguru = "^0.5.3"
tottle = "^1.4"
Jinja2 = "^3.0.1"
Markdown = "^3.3.4"


[tool.poetry.dev-dependencies]
howdoi = "^2.0.16"
isort = "^5.9.1"
black = "^21.6b0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
